# contracts/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/campaigns/{campaignId}/Encounter": {
      "get": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EncounterSummaryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EncounterSummaryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EncounterSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEncounterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEncounterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEncounterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaignId}/Encounter/{id}": {
      "get": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/campaigns/{campaignId}/Encounter/{id}/participants": {
      "post": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddParticipantRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddParticipantRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddParticipantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaignId}/Encounter/{id}/participants/{participantId}": {
      "delete": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaignId}/Encounter/{id}/participants/{participantId}/damage": {
      "post": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyEncounterDamageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyEncounterDamageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyEncounterDamageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaignId}/Encounter/{id}/participants/{participantId}/conditions/{conditionId}": {
      "post": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "conditionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "conditionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaignId}/Encounter/{id}/advance-turn": {
      "post": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaignId}/Encounter/{id}/next-round": {
      "post": {
        "tags": [
          "Encounter"
        ],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddParticipantRequest": {
        "type": "object",
        "properties": {
          "participantType": {
            "$ref": "#/components/schemas/ParticipantType"
          },
          "characterId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "monsterId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "initiative": {
            "type": "integer",
            "format": "int32"
          },
          "maxHP": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ac": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ApplyEncounterDamageRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "syncToCharacter": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CreateEncounterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "sessionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EncounterDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "campaignId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "round": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EncounterSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "round": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "participantCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ParticipantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "participantType": {
            "$ref": "#/components/schemas/ParticipantType"
          },
          "characterId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "monsterId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "initiative": {
            "type": "integer",
            "format": "int32"
          },
          "currentHP": {
            "type": "integer",
            "format": "int32"
          },
          "maxHP": {
            "type": "integer",
            "format": "int32"
          },
          "ac": {
            "type": "integer",
            "format": "int32"
          },
          "turnOrder": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "conditions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ParticipantType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  }
}